Fix dmcli command fail due to unavailable destination component
Some dmcli commands are issued way before destination component is
active. Added code for the routine when there is switching from
router to bridge and bridge to router mode. The command will be
issued till it succeeds with timeout of 30 secs.

Note that this only applies when there is a mode change, which is not
the regular boot up sequence.