retry dmcli commands while switching between bridge <-> router modes
Fix for dmcli commands failing due to unavailable destination component.

Some dmcli commands are issued before the destination component is
active. Added code for the routine when there is switching from
router to bridge and bridge to router mode. The command will be
retried until it succeeds or times out.

Note that this only applies when there is a mode change, which is not
the regular boot up sequence.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

add retry of LanMode update if failed during boot up provisioning ( OFW-3534 F3896SI-1416 )

fix Modem continues to be in bridge mode after switching to router mode.

BUG DESCRIPTION:

Modem continues to be in bridge mode after switching to router mode.

The system call used to set the value of Device.X_CISCO_COM_DeviceControl.LanManagementEntry.1.LanMode is failing Intermittently at the time of boot-up.

FIX DESCRIPTION:

Added changes to retry to set the dml if it's not succeeded in the first time.

TESTING PERFORMED:

1.Provision the modem with max cpe 2  in bridge mode using bootfile
2.Reboot from GUI
3.Provision the modem with max cpe 2 in router mode using bootfile
4.Reboot from GUI
5.Device changed to router mode and gpv of Device.X_CISCO_COM_DeviceControl.LanManagementEntry.1.LanMode is showing value as router

Upstream Candidate: May be, On top of TLV202 implementation.

Signed-off-by: Saravanakumar Karthikeyan <skarthikeyan.contractor@libertyglobal.com>