OFW-219: Modem mode DUT change is not persistent after reboot
Modified erouter mode based on the order of precendence according to the
spec and MV1 requirements.
1. eRouterSnmpInitMode [High]
2. webUI mode
3. CM config [Low]

signed-off-by XXXX <XXXX@YYYY.com>

fix switching to modem mode by setting TR-069 object ErouterModeControl to "ipDisabled" ( OFW-163 )

Function: void docsis_gotEnable_callback(Uint8 state)

This function was overriding the last set ErouterModeControl value
(through dmcli/ACS), at the time of boot-up. It has been always been
overriden with the boot config file's InitializationMode. This code
was introduced as part of OFW-219. This has been corrected. After the
code modification, ensured that the modem is booting-up in the
correct mode which was set earlier through ACS.

Function: int getDecisionErouteOperMode()

In honor mode 'eRouterMode' should not be overriden. Because
'eRouterMode' was getting over-written in the function and the honor
mode was not working. In Honor mode the DUT has to honor the mode
which is specified in the boot config file.

signed-off-by XXXX <XXXX@YYYY.com>

fix Device Mode is not Changing from Modem Mode to Ipv4/Dualstack from boot config file ( OFW-455 )

Steps:
1. Factory reset device .
2. Device is up and provisioned with mode defined in boot config file .
3. Change device mode to Modem Mode from boot config file and reboot device.
4. Device is changed to modem mode once box is up .
5. Change device mode to ipv4Mode from boot config file .
6. Once box is up device is in Modem Mode .
Expected Result : Device Mode Should be changed to ipv4Only after changing from boot config file .

Resolution:
At the time of boot up, GWP_Update_ErouterMode_by_InitMode()  will be called. Since the ErouterModeControl is honoreRouterInitMode, getDecisionErouteOperMode() changes eRouterMode to DOCESAFE_EROUTER_OPER_DISABLED_extIf from the proper value (cfg file value), which is not correct. So identified that the problem is happening in because of that code. Also compared with MV2 and identified that in MV2 is not having this changes. So made it compliance with MV2 implementation.

signed-off-by XXXX <XXXX@YYYY.com>