fix switching to modem mode by setting TR-069 object ErouterModeControl to "ipDisabled" ( OFW-163 )
Function: void docsis_gotEnable_callback(Uint8 state)

This function was overriding the last set ErouterModeControl value
(through dmcli/ACS), at the time of boot-up. It has been always been
overriden with the boot config file's InitializationMode. This code
was introduced as part of OFW-219. This has been corrected. After the
code modification, ensured that the modem is booting-up in the
correct mode which was set earlier through ACS.

Function: int getDecisionErouteOperMode()

In honor mode 'eRouterMode' should not be overriden. Because
'eRouterMode' was getting over-written in the function and the honor
mode was not working. In Honor mode the DUT has to honor the mode
which is specified in the boot config file.