fix Device Mode is not Changing from Modem Mode to Ipv4/Dualstack from boot config file ( OFW-455 )
Steps:
1. Factory reset device .
2. Device is up and provisioned with mode defined in boot config file .
3. Change device mode to Modem Mode from boot config file and reboot device.
4. Device is changed to modem mode once box is up .
5. Change device mode to ipv4Mode from boot config file .
6. Once box is up device is in Modem Mode .
Expected Result : Device Mode Should be changed to ipv4Only after changing from boot config file .

Resolution:
At the time of boot up, GWP_Update_ErouterMode_by_InitMode()  will be called. Since the ErouterModeControl is honoreRouterInitMode, getDecisionErouteOperMode() changes eRouterMode to DOCESAFE_EROUTER_OPER_DISABLED_extIf from the proper value (cfg file value), which is not correct. So identified that the problem is happening in because of that code. Also compared with MV2 and identified that in MV2 is not having this changes. So made it compliance with MV2 implementation.