optimise startup time ( OFW-870 )
Multiple entries in config file taking longer to reflect on DUT.

Optimize number of dmcli calls by sorting the entries and parsing the
output of getvalues for an object.

Note that this relies on a corresponding change in dmcli to add
support for a command line option to disable Alias Manager.