add retry of LanMode update if failed during boot up provisioning ( OFW-3534 F3896SI-1416 )
fix Modem continues to be in bridge mode after switching to router mode.

BUG DESCRIPTION:

Modem continues to be in bridge mode after switching to router mode.

The system call used to set the value of Device.X_CISCO_COM_DeviceControl.LanManagementEntry.1.LanMode is failing Intermittently at the time of boot-up.

FIX DESCRIPTION:

Added changes to retry to set the dml if it's not succeeded in the first time.

TESTING PERFORMED:

1.Provision the modem with max cpe 2  in bridge mode using bootfile
2.Reboot from GUI
3.Provision the modem with max cpe 2 in router mode using bootfile
4.Reboot from GUI
5.Device changed to router mode and gpv of Device.X_CISCO_COM_DeviceControl.LanManagementEntry.1.LanMode is showing value as router

Upstream Candidate: May be, On top of TLV202 implementation.

Signed-off-by: Saravanakumar Karthikeyan <skarthikeyan.contractor@libertyglobal.com>