fix zombie process is seen in gw_prov_utopia ( OFW-4298 )
[Issue]
Zombie process is observed in gw_prov_utopia during bootup.
This process is created because the popen function was not closed properly.

[Solution]
To resolve this issue, you need to perform a pclose after using popen.

[Testing Performed]
Checked all the process status after the bootup. No Zombie process was seen.

Upstreamable : Yes

Signed-off-by: Gokulakrishnan Ramalingam (gramalingam.contractor@libertyglobal.com)