fix ACS and RDK stop working ( OFW-4223 CH7465OFW-1305 )
[Bug description]
The rbus broker process (rtrouted) gets stuck after a few days in case
of docsis disconnection and reconnection.

The issue happens because gw_prov_utopia kills the thread that's
supposed to handle receiving rbus messages by mistake, and afterwards when
setting the values from the bootfile, requests are sent to rtrouted but
the responses are not handled because the rbus reader thread is dead.

gw_prov_utopia attempts to set each parameter 120 times before
continuing to the next one. After a certain number of rbus calls, the
socket send buffer on the rtrouted side towards gw_prov_utopia becomes
full and the send blocks which completely blocks rbus.

The rbus receive thread is killed because gw_prov_utopia calls
pthread_cancel to cancel a detached thread which was launched to start
hotspot, but by that time the thread has exited already and its ID was reused for the rbus
receiver thread.

[Solution]
Make the hotspot thread joignable instead of detached so that it's safe
to cancel it.

[Testing performed]
Boot up the gateway using the boot file from the ticket
After WAN is fully up, unplug the RF cable
After one minute, replug the RF cable
Wait five minutes, then unplug the RF cable
Wait one minute, then replug the RF cable
Check /rdklogs/logs/GWPROVLog.txt.* and verify that the error "Ccsp msg bus internal error" is not seen