remove duplicate entry from the failed list if the setv of current tlv parameter is success ( OFW-4829 )
[Issue]
[CSMV1-R1.6]-TLV202.1 taking precedence over TLV43.202.1 while added parameters in config file.

[Description]
While applying the tlv parameters in gwprovapp, the tr69 component is not up by that time.
So the first entry of tr69 paramater will throw error and it will keep in the failed list to reapply again once all parameters setv done.
Once all parameters setv done, the first parameter which failed will apply again.So its overwriting the last value with first value.

[Solution]
Remove the parameter from the failed list if the setv of current tlv parameter is success.

[Testing]
Verified the TLV specified at the end of config file have higher precedence.

[To Do]
This commit won't be part of MV2+ as the commit related to failed list is reverted ( OFW-4670 OFW-4688 ).

[Upstream Candidate]
Maybe

Signed-off-by: Sreeja Sreekumar (ssreekumar.contractor@libertyglobal.com)